+++ js/modules/controlSection.js
@@ -16,116 +16,7 @@
      */
     module.div = document.getElementById("ts3-control");
 
-    // Todo: doc
-    module.navbar = {
-
-        /**
-         * The element containing the navbar.
-         * @type {Element}
-         */
-        div: document.createElement("div"),
-
-        scrollToButtons: {},
-
-        /**
-         * Adds a scrollTo button to the navbar and adds the button element to the scrollToButtons object.
-         * The button has an onclick event that scrolls the table into view.
-         * Note: the button is hidden by default. (see showScrollToButton and hideScrollToButton)
-         * @param tableModule TODO
-         */
-        addScrollToButton: function (tableModule) {
-            var button = document.createElement("button");
-            this.scrollToButtons[tableModule.name] = button;
-            button.style.display = "none";
-
-            // Todo: uppercase scrollTo innerHTML
-            button.innerHTML = tableModule.name + "s";
-
-            button.onclick = function () {
-                tableModule.div.scrollIntoView();
-            };
-
-            this.div.appendChild(button);
-        },
-
-        showScrollToButton: function (tableModule) {
-            this.scrollToButtons[tableModule.name].style.display = "";
-        },
-
-        hideScrollToButton: function (tableModule) {
-            this.scrollToButtons[tableModule.name].style.display = "none";
-        }
-    };
-
-    // Todo: doc
-    module.tableSelection = {
-
-        /**
-         * The element containing the table selection section.
-         * @type {Element}
-         */
-        div: document.createElement("div"),
-
-        addTableCheckboxDiv: function (tableModule) {
-            var div = document.createElement("div");
-
-            var checkbox = document.createElement("input");
-            checkbox.id = tableModule.name + "Checkbox";
-            checkbox.type = "checkbox";
-            tableModule.checkbox = checkbox;
-
-            var label = document.createElement("label");
-            // Todo: proper innerHTML (e.g. 'Client table', currently is 'client')
-            label.innerHTML = tableModule.name;
-            label.htmlFor = tableModule.name + "Checkbox";
-
-            // Styling of the divs
-            switch (tableModule) {
-                case ts3ldv.tables.ban:
-                    div.className = "small-12 medium-6 large-6 columns";
-                    break;
-                case ts3ldv.tables.client:
-                    div.className = "small-12 medium-12 large-6 columns";
-                    break;
-
-                default:
-                    div.className = "small-12 medium-6 large-4 columns";
-            }
-
-            div.appendChild(checkbox);
-            div.appendChild(label);
-            this.div.appendChild(div);
-        },
-
-        /**
-         * TODO: doc
-         * Adds a checkbox listener for the table and displays/hides the navbar scroll buttons for the table.
-         *
-         * @param {string} table - name of the table (e.g. "clientTable"). TODO
-         */
-        addTableCheckboxEventListener: function (tableModule) {
-            tableModule.checkbox.onclick = function () {
-                if (this.checked) {
-                    ts3ldv.storage.setTableActive(tableModule, true);
-
-                    // Build table if necessary (= not already built and only hidden)
-                    if (!ts3ldv.storage.getTableBuilt(tableModule))
-                        ts3ldv.tables.buildTableEnhanced(tableModule);
-
-                    tableModule.div.style.display = "";
-                    module.navbar.showScrollToButton(tableModule);
-                } else {
-                    ts3ldv.storage.setTableActive(tableModule, false);
-                    tableModule.div.style.display = "none";
-                    module.navbar.hideScrollToButton(tableModule);
-                }
-
-                ts3ldv.tables.setFilterPlaceholders(tableModule);
-            }
-        }
-    };
-
-    // Todo: split into multiple (private?) functions (e.g. one for the navbar)
+    // Todo: split into multiple modules/functions
     /**
      * Builds the control section.
      */
@@ -232,27 +123,10 @@
 
         miscControlSection.appendChild(resetSortingButton);
 
-        var scrollBackToTopButton = document.createElement("button");
-
-        this.navbar.div.id = "navbar";
-        scrollBackToTopButton.innerHTML = "Top";
-
-        scrollBackToTopButton.onclick = function () {
-            scrollTo(0, 0);
-        };
-
-        this.navbar.div.appendChild(scrollBackToTopButton);
-        this.navbar.addScrollToButton(ts3ldv.tables.client);
-        this.navbar.addScrollToButton(ts3ldv.tables.ban);
-        this.navbar.addScrollToButton(ts3ldv.tables.kick);
-        this.navbar.addScrollToButton(ts3ldv.tables.complaint);
-        this.navbar.addScrollToButton(ts3ldv.tables.upload);
-
         controlSection.appendChild(this.tableSelection.div);
         controlSection.appendChild(creationTimestampSection);
         controlSection.appendChild(buildSection);
         controlSection.appendChild(miscControlSection);
-        controlSection.appendChild(this.navbar.div);
         this.div.appendChild(controlSection);
     };